+++ checkpoint/orbax/checkpoint/experimental/v1/_src/handlers/global_registration.py
@@ -23,6 +23,7 @@ given checkpointable will be used.
 from typing import Type
 
 from orbax.checkpoint.experimental.v1._src.handlers import json_handler
+from orbax.checkpoint.experimental.v1._src.handlers import leaf_handler
 from orbax.checkpoint.experimental.v1._src.handlers import proto_handler
 from orbax.checkpoint.experimental.v1._src.handlers import pytree_handler
 from orbax.checkpoint.experimental.v1._src.handlers import registration
@@ -54,3 +55,9 @@ _try_register_handler(pytree_handler.PyTreeHandler)
 _try_register_handler(
     pytree_handler.PyTreeHandler, format_utils.PYTREE_CHECKPOINTABLE_KEY
 )
+
+# Registration for leaf types that can be treated as distinct checkpointables.
+_try_register_handler(leaf_handler.ShardedArrayHandler)
+_try_register_handler(leaf_handler.ArrayHandler)
+_try_register_handler(leaf_handler.ScalarHandler)
+_try_register_handler(leaf_handler.StringHandler)